pyproject.toml
[project]
name = "voice-to-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg>=1.4",
    "google-generativeai>=0.8.5",
    "gradio>=5.27.1",
    "numpy>=2.2.5",
    "openai-whisper",
    "python-dotenv>=1.1.0",
    "transformers>=4.51.3",
    "whisper>=1.1.10",
]

[tool.uv.sources]
openai-whisper = { git = "https://github.com/openai/whisper.git" }
